# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Market API
  version: 1.0.0
  description: |
    Backend API cho **Smart Market – Chợ thông minh** (Zalo Mini App + CMS).

    ## Xác thực
    - **Mini app**: `POST /auth/zalo-login` với access token Zalo. Môi trường dev bật `ZALO_MOCK=true`
      cho phép token dạng `mock:<zaloId>:<Tên>` (vd `mock:zalo-buyer-1:Người Mua`).
    - **CMS**: `POST /auth/login` với username/email/số điện thoại + mật khẩu.
    - Sau khi đăng nhập, bấm **Authorize** và dán `accessToken` (không cần chữ "Bearer").

    ## Vai trò & phạm vi
    | Role | Phạm vi |
    |---|---|
    | super_admin | Toàn hệ thống |
    | province_admin | Các chợ trong tỉnh được gán (chủ yếu chỉ đọc + escalate phản ánh) |
    | market_manager | CRUD trong các chợ được gán |
    | user | Người mua; nếu `merchant_status=active` và có sạp → dùng được nhóm `/seller/*` |

    ## Quy ước
    - Envelope: `{ success, data, meta? }` hoặc `{ success: false, error: { code, message, details? } }`.
    - Phân trang: `?page=1&limit=20` (limit tối đa 100) → `meta { page, limit, total, totalPages }`.
    - Tiền VND số nguyên. Thời gian ISO 8601. Key JSON camelCase.
    - Tài khoản demo (seed): SA `admin@smartmarket.vn / Admin@12345`; MM `mm.trungtam@smartmarket.vn / Manager@123`.
servers:
  - url: http://localhost:3000/api/v1
    description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Search:
      name: search
      in: query
      schema: { type: string }
    MarketIdQuery:
      name: marketId
      in: query
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: STATE_INVALID }
            message: { type: string, example: Đơn đang ở trạng thái "completed", không thể chuyển sang "confirmed" }
            details: { nullable: true }
    Image:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri }
    ProductTraceabilityInput:
      type: object
      required: [producerName]
      properties:
        producerName: { type: string, maxLength: 255, description: Đơn vị sản xuất hoặc cung cấp }
        productionAddress: { type: string, nullable: true, description: Địa chỉ hoặc vùng sản xuất }
        batchCode: { type: string, maxLength: 100, nullable: true }
        productionDate: { type: string, format: date, nullable: true }
        harvestDate: { type: string, format: date, nullable: true }
        expiryDate: { type: string, format: date, nullable: true }
        certificateName: { type: string, maxLength: 255, nullable: true }
        certificateNumber: { type: string, maxLength: 100, nullable: true }
        documents:
          type: array
          maxItems: 6
          items: { $ref: '#/components/schemas/Image' }
        externalUrl: { type: string, format: uri, nullable: true }
        notes: { type: string, maxLength: 5000, nullable: true }
...
  responses:
    Ok:
      description: Thành công
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: {}
    Paginated:
      description: Danh sách có phân trang
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: true }
              data: { type: array, items: {} }
              meta:
                type: object
                properties:
                  page: { type: integer, example: 1 }
                  limit: { type: integer, example: 20 }
                  total: { type: integer, example: 135 }
                  totalPages: { type: integer, example: 7 }
    Err:
      description: Lỗi (400/401/403/404/409/429 — xem `error.code`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security: []
paths:
...
  /products:
    get:
      tags: [Products (App)]
      summary: Danh sách sản phẩm theo chợ (kèm finalPrice, stockStatus, hasTraceability)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: marketId, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: categoryId, in: query, schema: { type: string, format: uuid } }
        - { name: stallId, in: query, schema: { type: string, format: uuid } }
        - { name: onSale, in: query, schema: { type: boolean } }
        - { name: inStock, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string, enum: [price, rating, newest] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
  /products/{id}:
    get:
      tags: [Products (App)]
      summary: Chi tiết sản phẩm + hồ sơ truy xuất tự khai báo + tiểu thương + sản phẩm liên quan
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /products/{id}/reviews:
    get:
      tags: [Products (App)]
      summary: Đánh giá của sản phẩm
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }

  # ================= ORDERS (BUYER) =================
...
  /seller/products:
    get:
      tags: [Seller]
      summary: Sản phẩm của sạp (tab all/selling/hidden/out_of_stock/on_sale)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: tab, in: query, schema: { type: string, enum: [all, selling, hidden, out_of_stock, on_sale] } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
    post:
      tags: [Seller]
      summary: Tạo sản phẩm
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, images, categoryId, price, unit, quantity]
              properties:
                name: { type: string }
                images: { type: array, minItems: 1, items: { $ref: '#/components/schemas/Image' } }
                categoryId: { type: string, format: uuid }
                price: { type: integer, minimum: 1, example: 25000 }
                unit: { type: string, example: kg, description: Thuộc danh mục product_units trong settings }
                quantity: { type: number, example: 50 }
                origin: { type: string, nullable: true }
                description: { type: string, nullable: true }
                isHidden: { type: boolean, default: false }
                traceability:
                  allOf: [{ $ref: '#/components/schemas/ProductTraceabilityInput' }]
                  nullable: true
      responses:
        '201': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
  /seller/products/{id}:
    get:
      tags: [Seller]
      summary: Chi tiết sản phẩm của sạp
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
    put:
      tags: [Seller]
      summary: Sửa sản phẩm
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                images: { type: array, items: { $ref: '#/components/schemas/Image' } }
                categoryId: { type: string, format: uuid }
                price: { type: integer }
                unit: { type: string }
                quantity: { type: number }
                origin: { type: string }
                description: { type: string }
                isHidden: { type: boolean }
                traceability:
                  allOf: [{ $ref: '#/components/schemas/ProductTraceabilityInput' }]
                  nullable: true
                  description: Gửi null để xóa hồ sơ truy xuất của sản phẩm
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }
    delete:
      tags: [Seller]
      summary: Xóa sản phẩm (soft delete)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /seller/products/{id}/visibility:
    patch:
      tags: [Seller]
      summary: Ẩn/hiện sản phẩm
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isHidden]
              properties:
                isHidden: { type: boolean }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /admin/products:
    get:
      tags: [Admin - Products]
      summary: Giám sát sản phẩm theo scope
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MarketIdQuery'
        - { name: stallId, in: query, schema: { type: string, format: uuid } }
        - { name: includeHidden, in: query, schema: { type: boolean, default: true } }
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200': { $ref: '#/components/responses/Paginated' }
  /admin/products/{id}/visibility:
    patch:
      tags: [Admin - Products]
      summary: Ẩn/hiện sản phẩm vi phạm (thông báo cho tiểu thương)
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isHidden, reason]
              properties:
                isHidden: { type: boolean }
                reason: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        default: { $ref: '#/components/responses/Err' }

  # ================= ADMIN - ORDERS =================
...